Fixed most of the lag caused by autoplay
Autoplay checks to see if the current song is finished. Having it check every loop was excessive so now it only checks once every 100 loops. arrow keys are more responsive and the mouse wheel is improved but still not perfect.
playing status indicator also functions properly if mocs is opened while a song is already playing