replaced more subprocess cmds with socket cmds
songs are now added to the queue via the mocp socket instead of being added to the playlist with the subprocess command. This doesn't fix the slow frozen ui bug for now.

also fixed a window resize bug by ensuring that the list of songs on screen matches the window height. replaced tuple with list.

tryit decorator should now return value of decorated function.